Removed test case.
After the changes to it, this test doesn't test any more what it was supposed to test (that implicit conversion is done to first convertible type and not first type that can hold the value). So it can be removed now because it tests the same thing as the test above it.