Rip out custom select_type in favor of tuple_element
This rips out the custom select type implementation in favor of the
C++11 conforming std::tuple_element. Note that std::tuple_element is
already included by the utility header, which includes the tuple header
itself. We only add the tuple header to be on the safe side here.

I made sure the tests via `make test` work; from `make bench` I can not
really see a difference, as the values are so small.