Implements Pattern Matching for Sum Types via `.match` Member Function.
Member function `.match` for accepting an arbitrary number of lambdas
building a single functor out of the lambdas for visitation.

This mimics pattern matching as known from languages like Haskell,
and works with C++ lambdas for convenience. In fact lambdas are the
primary if not only use-case for this at the moment.

Usage:

    variant<T, U> v;

    v.match([](T){ cout << "I'm a T"; },
            [](U){ cout << "I'm a U"; });

Note: you can use lambda captures (e.g. `[&](T){ ts += 1; }`) to
interact with the outside easily.

References:
- https://www.haskell.org/definition/haskell2010.pdf 3.17
- https://doc.rust-lang.org/book/enums.html
- https://doc.rust-lang.org/book/match.html
- http://dl.acm.org/citation.cfm?id=2384686
- http://stroustrup.com/OOPSLA-typeswitch-draft.pdf