Promotes Variant to mapbox:: namespace, closes #104.
In a future version we want to remove the `mapbox::util::`
namespace. Until then we provide both adding a deprecation
warning to the readme.

Implementation detail: we make the util namespace an `inline`
namespace - otherwise users could notice us changing the
specific namespace in which our templates were declared in.

The downside here is we can't add a deprecation pragma since
then the `mapbox::` namespace would contain it, too.